Fixed a bug in chbg.sh
Turns out, I didn't consider word splitting when selecting backgrounds
in random_shuffle select mode which led to some file names(the
background paths with space) to be split and hence invalid file names.
More to that, I've updated my bashrc by activating nullglob, failglob
and globstar which have turned out to be very important in globbing as
I've learned the hard way when running `nvim *.txy` in which nvim opens
a new file `*.txy` since no file ends with that extension which is
clearly an error to `*.txt`, it should fail with an error instead.